fix: extract mask parsing from CurrentMaskOverlay to OverlayEvents
